Typescriptify the Dataset and Snapshot table components
